requirements.txt
kafka-python==2.0.2
avro==1.11.3
requests